refactor: basic context manager for MontyExperiment (#197)
Not using a context manager means if an exception is thrown before
`close` is called, the dataset isn't closed. In some environments, we're
getting KeyErrors from one of the tests and this leads to crashing
workers in other tests.

This change adds some basic context manager methods to MontyExperiment
and updates its uses in the tests to use `with` expressions to ensure
that `close` is called correctly. This is only a first step though, and
we should probably consider some more changes to make MontyExperiment
use the context manager pattern more fully.